==> 01_prepacacao_dados_pandas.ipynb <==
## Preparação dos arquivos CSV para tipos compatíveis com o Schema

import pandas as pd
import numpy as np
from pathlib import Path

pd.set_option('display.max_rows', 500)

origin = Path('raw_data')
destiny = Path('data')

#### Person

df = pd.read_csv(origin/'Person.Person.csv', sep=';', na_values=['NULL'])

df.to_csv(destiny/'Person.Person.csv', index=False, sep=';')

#### Product

df = pd.read_csv(origin/'Production.Product.csv', sep=';', na_values=['NULL'])

df['ProductSubcategoryID'] = df['ProductSubcategoryID'].fillna(-1)
df['ProductSubcategoryID'] = df['ProductSubcategoryID'].astype(int)
df['ProductSubcategoryID'] = df['ProductSubcategoryID'].astype(str)
df['ProductSubcategoryID'] = df['ProductSubcategoryID'].replace('-1', np.nan)

df['ProductModelID'] = df['ProductModelID'].fillna(-1)
df['ProductModelID'] = df['ProductModelID'].astype(int)
df['ProductModelID'] = df['ProductModelID'].astype(str)
df['ProductModelID'] = df['ProductModelID'].replace('-1', np.nan)

df['StandardCost'] = df['StandardCost'].str.replace(',', '.')
df['ListPrice'] = df['ListPrice'].str.replace(',', '.')

df.to_csv(destiny/'Production.Product.csv', index=False, sep=';')

#### Customer

df = pd.read_csv(origin/'Sales.Customer.csv', sep=';', na_values=['NULL'])

df['PersonID'] = df['PersonID'].fillna(-1)
df['PersonID'] = df['PersonID'].astype(int)
df['PersonID'] = df['PersonID'].astype(str)
df['PersonID'] = df['PersonID'].replace('-1', np.nan)

df['StoreID'] = df['StoreID'].fillna(-1)
df['StoreID'] = df['StoreID'].astype(int)
df['StoreID'] = df['StoreID'].astype(str)
df['StoreID'] = df['StoreID'].replace('-1', np.nan)

df.to_csv(destiny/'Sales.Customer.csv', index=False, sep=';')

#### SalesOrderHeader

df = pd.read_csv(origin/'Sales.SalesOrderHeader.csv', sep=';', na_values=['NULL'])

df['SalesPersonID'] = df['SalesPersonID'].fillna(-1)
df['SalesPersonID'] = df['SalesPersonID'].astype(int)
df['SalesPersonID'] = df['SalesPersonID'].astype(str)
df['SalesPersonID'] = df['SalesPersonID'].replace('-1', np.nan)

df['CreditCardID'] = df['CreditCardID'].fillna(-1)
df['CreditCardID'] = df['CreditCardID'].astype(int)
df['CreditCardID'] = df['CreditCardID'].astype(str)
df['CreditCardID'] = df['CreditCardID'].replace('-1', np.nan)

df['CurrencyRateID'] = df['CurrencyRateID'].fillna(-1)
df['CurrencyRateID'] = df['CurrencyRateID'].astype(int)
df['CurrencyRateID'] = df['CurrencyRateID'].astype(str)
df['CurrencyRateID'] = df['CurrencyRateID'].replace('-1', np.nan)

df['SubTotal'] = df['SubTotal'].str.replace(',', '.')
df['TaxAmt'] = df['TaxAmt'].str.replace(',', '.')
df['Freight'] = df['Freight'].str.replace(',', '.')
df['TotalDue'] = df['TotalDue'].str.replace(',', '.')

df.to_csv(destiny/'Sales.SalesOrderHeader.csv', index=False, sep=';')

#### SpecialOfferProduct

df = pd.read_csv(origin/'Sales.SpecialOfferProduct.csv', sep=';', na_values=['NULL'])

df.to_csv(destiny/'Sales.SpecialOfferProduct.csv', index=False, sep=';')

#### SalesOrderDetail

df = pd.read_csv(origin/'Sales.SalesOrderDetail.csv', sep=';', na_values=['NULL'])

df['UnitPrice'] = df['UnitPrice'].str.replace(',', '.')
df['UnitPriceDiscount'] = df['UnitPriceDiscount'].str.replace(',', '.')

df.to_csv(destiny/'Sales.SalesOrderDetail.csv', index=False, sep=';')
