==> 02_envio_arquivos_s3.ipynb <==
## Envio dos arquivos CSV para o S3 usando client Python

import pandas as pd
import boto3
import json
import logging

from pathlib import Path
from botocore.exceptions import ClientError

import warnings
warnings.filterwarnings('ignore')

### Carregar parametros do arquivo de configuração

import configparser
config = configparser.ConfigParser()
config.read_file(open('dwh.cfg'))

KEY                    = config.get('AWS','KEY')
SECRET                 = config.get('AWS','SECRET')

### Criar cliente S3

s3 = boto3.resource('s3',
                       region_name="us-east-1",
                       aws_access_key_id=KEY,
                       aws_secret_access_key=SECRET
                   )

### Fazer upload dos arquivos para o bucket S3

path = Path('data')
files = [str(file) for file in path.iterdir()]

files

def upload_file(file_name, bucket, object_name=None):
    """Upload a file to an S3 bucket

    :param file_name: File to upload
    :param bucket: Bucket to upload to
    :param object_name: S3 object name. If not specified then file_name is used
    :return: True if file was uploaded, else False
    """

    # If S3 object_name was not specified, use file_name
    if object_name is None:
        object_name = file_name

    # Upload the file
    s3_client = boto3.client('s3')
    try:
        response = s3_client.upload_file(file_name, bucket, object_name)
    except ClientError as e:
        logging.error(e)
        return False
    return True

# https://boto3.amazonaws.com/v1/documentation/api/latest/guide/s3-uploading-files.html

bucket = 'roxproject'

for file_name in files:    
    upload_file(file_name, bucket)

roxProjectBucket = s3.Bucket(bucket)
for obj in roxProjectBucket.objects.all():
    print(obj)
