==> 04_analise_dados_spark.ipynb <==
import findspark

findspark.init()

from pyspark.sql import SparkSession, SQLContext
from pyspark.sql.functions import regexp_replace

spark = SparkSession.builder.appName('roxproject').getOrCreate()
sc =  spark.sparkContext
sqlContext = SQLContext(sc)

df_person = spark.read.csv('data/Person.Person.csv', 
                            sep=';', inferSchema=True, header=True)

df_person.createOrReplaceTempView('Person')

df_product = spark.read.csv('data/Production.Product.csv', 
                            sep=';', inferSchema=True, header=True)

df_product.createOrReplaceTempView('Product')

df_costumer = spark.read.csv('data/Sales.Customer.csv', 
                            sep=';', inferSchema=True, header=True)

df_costumer.createOrReplaceTempView('Customer')

df_orderdetail = spark.read.csv('data/Sales.SalesOrderDetail.csv', 
                            sep=';', inferSchema=True, header=True)

df_orderdetail.createOrReplaceTempView('SalesOrderDetail')

df_orderheader = spark.read.csv('data/Sales.SalesOrderHeader.csv', 
                            sep=';', inferSchema=True, header=True)

df_orderheader = df_orderheader.withColumn('TotalDue', regexp_replace('TotalDue', ',', '.'))
df_orderheader = df_orderheader.withColumn('TotalDue', df_orderheader['TotalDue'].cast("float"))

df_orderheader.createOrReplaceTempView('SalesOrderHeader')

df_specialoffer = spark.read.csv('data/Sales.SpecialOfferProduct.csv', 
                            sep=';', inferSchema=True, header=True)

df_specialoffer.createOrReplaceTempView('SpecialOfferProduct')

### Análise de Dados

1. Escreva uma query que retorna a quantidade de linhas na tabela Sales.SalesOrderDetail pelo campo SalesOrderID, desde que tenham pelo menos três linhas de detalhes.
    

query = """
    SELECT COUNT(SalesOrderID) AS OrderQty
    FROM (
        SELECT SalesOrderID, COUNT(SalesOrderID) AS Qty
        FROM SalesOrderDetail
        GROUP BY SalesOrderID
    )
    WHERE Qty > 2
"""

sqlContext.sql(query).show(100, False)

2. Escreva uma query que ligue as tabelas Sales.SalesOrderDetail, Sales.SpecialOfferProduct e Production.Product e retorne os 3 produtos (Name) mais vendidos (pela soma de OrderQty), agrupados pelo número de dias para manufatura (DaysToManufacture).

query = """
    SELECT Name, SUM(OrderQty) AS OrderQtyTotal, DaysToManufacture
    FROM Product AS P
    JOIN SalesOrderDetail AS D ON D.ProductID = P.ProductID
    JOIN SpecialOfferProduct AS S ON S.ProductID = P.ProductID 
        AND S.SpecialOfferID = D.SpecialOfferID
    GROUP BY Name, DaysToManufacture
    ORDER BY OrderQtyTotal DESC
    LIMIT 3
"""

sqlContext.sql(query).show(100, False)

3. Escreva uma query ligando as tabelas Person.Person, Sales.Customer e Sales.SalesOrderHeader de forma a obter uma lista de nomes de clientes e uma contagem de pedidos efetuados.

query = """
    SELECT CONCAT_WS(' ', FirstName, NULLIF(MiddleName, 'NULL'), LastName) AS Name,
        COUNT(SalesOrderID) AS OrderQty
    FROM Person AS P
    JOIN Customer AS C ON C.PersonID = P.BusinessEntityID 
    JOIN SalesOrderHeader AS S ON C.CustomerID = S.CustomerID
    GROUP BY Name
    ORDER BY OrderQty DESC
"""

sqlContext.sql(query).show(20, False)

4. Escreva uma query usando as tabelas Sales.SalesOrderHeader, Sales.SalesOrderDetail e Production.Product, de forma a obter a soma total de produtos (OrderQty) por ProductID e OrderDate.
    

query = """
    SELECT P.ProductID, OrderDate, SUM(OrderQty) AS OrderQtyTotal
    FROM Product AS P
    JOIN SalesOrderDetail AS D ON D.ProductID = P.ProductID
    JOIN SalesOrderHeader AS S ON D.SalesOrderID = S.SalesOrderID
    GROUP BY P.ProductID, OrderDate
    ORDER BY OrderQtyTotal DESC
"""

sqlContext.sql(query).show(20, False)

5. Escreva uma query mostrando os campos SalesOrderID, OrderDate e TotalDue da tabela Sales.SalesOrderHeader. Obtenha apenas as linhas onde a ordem tenha sido feita durante o mês de setembro/2011 e o total devido esteja acima de 1.000. Ordene pelo total devido decrescente.

query = """
    SELECT SalesOrderID, OrderDate, TotalDue
    FROM SalesOrderHeader
    WHERE OrderDate >= '2011-09-01 00:00:00'
    AND OrderDate < '2011-10-01 00:00:00'
    AND TotalDue > 1000
    ORDER BY TotalDue DESC
"""

sqlContext.sql(query).show(20, False)
